src/pyproject.toml
[project]
name = "quartapp"
version = "1.0.0"
description = "Create a simple chat app using Quart and OpenAI"
dependencies = [
    "quart",
    "werkzeug",
    "gunicorn",
    "uvicorn[standard]",
    "openai>=1.108.1",
    "azure-identity",
    "aiohttp",
    "python-dotenv",
    "pyyaml",
    "azure-storage-blob>=12.19.0",
    "opencv-python-headless>=4.9.0.80",
    "numpy>=1.24.0"
    ]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"
